pyproject.toml
[tool.poetry]
name = "orchard-bot"
version = "0.1.0"
description = ""
authors = ["auburn <auburnete@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.15.0"
uvicorn = "^0.14.0"
python-dotenv = "^0.18.0"
PyNaCl = "^1.4.0"
httpx = "^0.18.2"

[tool.poetry.dev-dependencies]
ipython = "^7.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
